Separate bech32 implementation from payments code
